sm8250-common: Add uclamp support for kona
Thanks to task_profiles, subgroups of cpuctl (uclamp) were sorted and
populated. But the group of system-background is unexpectedly empty.
That's not good for us as some important system tasks are in cpuset but
not in cpuctl meaning that we can't set util bounds for them.

By json rules, this workaround accomplishs that all tasks in sys-bg of
cpuset should go and join the same group of cpuctl, except of already
grouped tasks. Now, we can set uclamp caps on system process.